+++ NEWS.md
@@ -1,4 +1,4 @@
-# systemfonts (development version)
+# systemfonts 1.1.0
 
 * `match_fonts()` have been added as a vectorized and generalized version of 
   `match_font()`. In the process `match_font()` has been deprecated in favour of